+++ examples/provider/provider.tf
@@ -8,6 +8,7 @@ terraform {
 
 provider "host" {
   target_user = "dongho"
+  aur_helper  = "yay"
 
   # runtime_dir defaults to ~/.local/state/terraform-provider-host for this user.
 }